requirements.txt
Django==3.2.13
djangorestframework==3.13.1
youtube-dl==2021.12.17
